Finishing with kernel.response and kernel.exception Our controller returns a Response, so we skip the entire kernel.view block and go straight down to the filterResponse() call:
// vendor/symfony/symfony/src/Symfony/Component/HttpKernel/HttpKernel.php
// ...

private function handleRaw(Request {dollar_sign}request, {dollar_sign}type = self::MASTER_REQUEST){
  // ...

  return {dollar_sign}this->filterResponse({dollar_sign}response, {dollar_sign}request, {dollar_sign}type);
}
This function shows up in one other place further up. If one of our kernel.request listeners sets the response, the response also goes through this function. So no matter how we create the response, filterResponse() is called.

The kernel.response Event

What does it do? Come on, you should be able to guess by now. It dispatches yet another event, and this time, it’s called kernel.response:

// vendor/symfony/symfony/src/Symfony/Component/HttpKernel/HttpKernel.php
// ...

private function filterResponse(Response {dollar_response}response, Request {dollar_response}request, {dollar_response}type){
  {dollar_response}event = new FilterResponseEvent({dollar_response}this, {dollar_response}request, {dollar_response}type, {dollar_response}response);

  {dollar_response}this->dispatcher->dispatch(KernelEvents::RESPONSE, {dollar_response}event);

  {dollar_response}this->finishRequest({dollar_response}request, {dollar_response}type);

  return {dollar_response}event->getResponse();
}
What’s awesome about kernel.response? Listeners to it have access to the Response object by calling getResponse() on the FilterResponseEvent object. So if you want to modify the response, like adding a header, this is your event.

How the Web Debug Toolbar Works

In fact, the web debug toolbar works via a listener to this event. When we load up a page, the web debug toolbar shows up at the bottom. This works because at the bottom of the HTML source, there’s a bunch of JavaScript that makes an AJAX request that loads it. But how does this JavaScript get there? The answer: with a listener on the kernel.response event. That listener injects the extra JavaScript into the HTML code of our page when we’re in development mode.

So all filterResponse() does is give us another hook point. It also calls finishRequest(), which is honestly less important:

// vendor/symfony/symfony/src/Symfony/Component/HttpKernel/HttpKernel.php
// ...

private function finishRequest(Request {dollar_sign}request, {dollar_sign}type){
  {dollar_sign}this->dispatcher->dispatch(KernelEvents::FINISH_REQUEST, new FinishRequestEvent({dollar_sign}this, {dollar_sign}request, {dollar_sign}type));
  {dollar_sign}this->requestStack->pop();
}
It dispatches another event and removes our request from the request stack. The request stack is an object that keeps track of all the requests we’re processing. We’ll talk about subrequests in a second, and then it’ll make sense how Symfony can be handling multiple requests at once.

After all this, the Response object is returned all the way back to app_dev.php. We made it through Symfony’s core and came back out alive with the response in hand:

// web/app_dev.php
// ...

{dollar_sign}response = {dollar_sign}kernel->handle({dollar_sign}request);
{dollar_sign}response->send();
{dollar_sign}kernel->terminate({dollar_sign}request, {dollar_sign}response);
What do we do with it? We call send(). This sends all the headers with the header() function and echo’s the content.