Simple: Event, Controller Event
Call me crazy, but when we zoom out, I think the request-response flow for Symfony is pretty darn simple. It’s basically: event, controller, event.

Go back and look at the Timeline in the profiler, because now, it tells a beautiful story. At the top, we see that kernel.request happens first, and everything below its bar are listeners. Then, it figures out which controller we want - that’s the controller.get_callable part. Cool. Next, it dispatches kernel.controller, and you can see its listeners. After that, the controller is called and the stuff under that is our work. We can see some Doctrine calls we’re making and the time it takes to render the template. After the controller, the kernel.response event is dispatched, it has a few listeners, and there’s kernel.terminate. Brilliant!

So guys, it’s just events, call the controller, then more events. Yep, that’s it. And now that we’ve journeyed to the core of Symfony’s request and response flow, let’s bend this to our will to do some crazy, custom things.