If this were the Matrix, you’d be seeing 1’s and 0’s running across your code. Your Symfony world is now one without rules or controls, without borders or boundaries. Let’s prove it.
We know that Symfony reads the _controller key in the request attributes and executes that as the controller. This is set by the RouterListener but it could be set anywhere.

Let’s go adventuring... UserAgentSubscriber listens to kernel.request. What if we just replaced the _controller key here with something else? Let’s do that, and set it to an anonymous function.

Symfony doesn’t see any difference between this function and a traditional controller. To prove it, return a normal Response:
return new Response('Hello world!')

When we refresh, would you believe it? That actually works!