Here’s our next challenge: pretend that it’s really important to know if the user is on a Linux or not. We’re really wanting to start measuring how hipster our user base is. In fact, it’s so important, that we want to be able to have an  argument in any controller function anywhere in the system. This won’t come from a routing wildcard like normal - we’ll figure this out by reading the User-Agent.
We’ll pass that into our template, and inside there, use it to print out a threatening message if the user is on a Linux.
Back to the homepage! If we try this now, we know what’s going to happen. We get a huge error because there is no {isLinux} in the route. Symfony has no idea what to pass to the argument. And up until now that’s been the rule: whatever we have in our routing curly brace is available as an argument and there are no exceptions to that rule, except for the Request object.