Allowing Other Controller Arguments
But guys, that’s not true! And we know it. We know that it’s not really about the routing layer. The arguments to the controller come from the request attributes. And sure, the only thing that normally modifies those is the routing layer. But there’s nothing stopping someone else from adding some extra stuff.

In our subscriber, let’s first get an  variable. We’ll look for the User-Agent header and we’ll look for the word Linux. Check to see if stripos doesn’t equal false.

To make this available as an argument, all we need to do is put it in the request attributes.

Seriously, that’s it. When we refresh, it works!