The Person Behind the Curtain (Event Listeners)
So why is this important? Because understanding the core of Symfony is letting you do things that previously looked impossible. You’re also going to be able to figure out how magic from outside libraries is working.

For example look at the SensioFrameworkExtraBundle(http://symfony.com/doc/current/bundles/SensioFrameworkExtraBundle/index.html). This is basically a bundle of shortcuts that work via magic. Now that you’ve journeyed to the center of Symfony and back, if you look at each shortcut, you should be able to explain the magic behind each of these.

The one I want to look at now is the ParamConverter(http://symfony.com/doc/current/bundles/SensioFrameworkExtraBundle/annotations/converters.html):

/**
 * @Route('/blog/{id}')
 * @ParamConverter('post', class='SensioBlogBundle:Post')
 */
public function showAction(Post {dollar_sign}post)
{
}
In the example, you can see that the controller has a  argument, but also that there’s no {post} in the routing. So this should throw an error. The ParamConverter, which works via a listener to kernel.controller, grabs the id off of the request attributes, queries for a Post object via Doctrine with that id, and then adds a new request attribute called post that’s set to that object:

// Summarized version of ParamConverterListener
public function onKernelController(FilterControllerEvent {dollar_sign}event) {
  {dollar_sign}request = {dollar_sign}event->getRequest();
  {dollar_sign}id = {dollar_sign}request->attributes->get('id');

  {dollar_sign}entity = {dollar_sign}this->em->getRepository('SensioBlogBundle:Post')
    ->find({dollar_sign}id);
  if (! {dollar_sign}entity) {
    throw new NotFoundHttpException('No Post found for '.{dollar_sign}id);
  }

  {dollar_sign}request->attributes->set('post', );
}
And just by doing that, the showAction can have that {dollar_sign}post argument.

If that makes any sense at all, you’re on the verge of really mastering a big part of Symfony.

Before we talk about sub requests, I want to point something out. If you’re playing with things, inside the profiler, there is a Request tab, which is interesting because it shows you the request attributes. You can see the _controller, the routes stuff and the isLinux key. By the way, not that it’s necessarily useful, but the fact that there is an _route key does mean that you can have a {dollar_sign}_route argument to any controller.