Have you ever been in a Twig template, and you suddenly need access to a variable you don’t have? Open up base.html.twig, and let’s pretend that we want to render the latest tweets from our Twitter account at the bottom.
Since there’s no one controller that fuels this template, there’s no way for us to pass the latest tweets to this template. When you’re in this spot, there are 2 fixes. First, you could create a new Twig function in a Twig extension. That’s usually the best approach. The second option is with the Twig render() function, which is our gateway to sub requests! This will also let you cache this chunk of HTML, but we’ll talk about that another time.

When we use render(controller('...')), this lets us execute a totally different controller. Let’s render a new controller called _latestTweetsAction in our same controller class.

I put the underscore in front of the name just as a reminder to myself that the controller only returns a fragment of HTML, not a full page.

In the controller, create the public function _latestTweetsAction. I don’t really want to bother with Twitter’s API right this second, so I’ll copy in some fake tweets instead. Now let’s render a template and pass those in.

Yea, this is only going to return a page fragment, we still always return a Response object from a controller. And that’s what we’re doing here. Now, let’s create this template, _latestTweets.html.twig. I’ll just paste some code in that loops over these tweets and puts them in a list.

Nothing scary here!

The Twig render() function will call out to that function, get the Response, grab its content, then put it at the bottom of the page. Scroll down - there it is!

OMG: You just made a Sub-Request!

Why did I show this? Why is this important to understanding the core? Open up the profiler and go back to the Timeline. Hmm, it looks normal at first and we can see where our controller and template are run. But hey, what’s that darker bar behind the template? Scroll way down to find a second request called a sub-request. OMG!

The sub-request is totally independent: it goes through the entire same process we just learned. It executes the kernel.request listeners, it dispatches the kernel.controller event, it calls the _latestTweetsAction controller and has kernel.response on the end. It really is like we are handling two totally separate request-response cycles.

Heck, you can even click to see the profiler for just the sub-request.