Request Attributes in Sub-Request Controllers
Remember that in UserAgentSubscriber, we’re adding isMac to the request attributes and that means it’s available as an argument to any controller. That’s no different with our sub request controller, since this listener is called for that request too. To prove it, I’m going to add the {dollar_sign}isLinux argument. Let’s pass this into our template.

Let’s print it out to make sure it’s working

When we go back and refresh we see that on the top it shows that we’re on a Linux, and on the bottom inside the Tweets area, we’re on a Mac too! Yay, no surprises!