A Disturbance in the Request
Here is where things get crazy. Go back to UserAgentSubscriber. Let’s add an override so it’s easier for us to play with this 'is Linuxs' stuff, since I’m pretty permanently using one.

If there’s a query parameter, called notLinux, that’s set to some value like 1, then let’s always set {dollar_sign}isLinux to false.

Back on the browser, when I refresh, I’m still on a Linux. But if I add a ?notLinux=1 to the URL, it goes away. The override correctly makes it look like I’m not on a Linux.

Now scroll down. Woh! The sub request still thinks we’re on a Linux. Something just short circuited in the system. But before we fix it, let’s dive one level deeper and see how sub-requests really work.