Passing Information to a Sub-Request Controller
But wait! I do want to know if the user is on a Mac from my sub request. What’s the solution?

The answer is really simple: just pass it to the controller. The second argument to the controller() function is an array of items that you want to make available as arguments to the controller. Behind the scenes, these are put onto the attributes of the sub request. So we can add a userOnLinux key and set its value to the true isLinux attribute stored on the master request. So, app.request.attributes.get('isLinux').

Inside of the controller, add a userOnLinux variable and pass it into the template.

Now when we refresh, we still have the ?notMac=1, so the Linux message is missing from the master request part at the top. And if we scroll down, the sub request also knows we’re not on a Linux because we’re passing that information through.

When we take off the query parameter, it looks like we’re on a mac up top on the bottom. Brilliant!

The lesson is that you need to be careful not to read outside request information, like query parameters from the URL, from inside a sub-request. This also ties into Http caching and ESI which are topics we’ll cover later. If we follow this rule and you do want to cache this latest tweets fragment, it’s going to be super easy.